+++ banyand/metadata/client.go
@@ -78,7 +78,7 @@ func (s *clientService) SchemaRegistry() schema.Registry {
 func (s *clientService) FlagSet() *run.FlagSet {
 	fs := run.NewFlagSet("metadata")
 	fs.StringVar(&s.namespace, "namespace", DefaultNamespace, "The namespace of the metadata stored in etcd")
-	fs.StringArrayVar(&s.endpoints, FlagEtcdEndpointsName, []string{"http://localhost:2379"}, "A comma-delimited list of etcd endpoints")
+	fs.StringSliceVar(&s.endpoints, FlagEtcdEndpointsName, []string{"http://localhost:2379"}, "A comma-delimited list of etcd endpoints")
 	fs.StringVar(&s.etcdUsername, flagEtcdUsername, "", "A username of etcd")
 	fs.StringVar(&s.etcdPassword, flagEtcdPassword, "", "A password of etcd user")
 	fs.StringVar(&s.etcdTLSCAFile, flagEtcdTLSCAFile, "", "Trusted certificate authority")
@@ -150,8 +150,10 @@ func (s *clientService) Serve() run.StopNotify {
 func (s *clientService) GracefulStop() {
 	s.closer.Done()
 	s.closer.CloseThenWait()
-	if err := s.schemaRegistry.Close(); err != nil {
-		logger.GetLogger(s.Name()).Error().Err(err).Msg("failed to close schema registry")
+	if s.schemaRegistry != nil {
+		if err := s.schemaRegistry.Close(); err != nil {
+			logger.GetLogger(s.Name()).Error().Err(err).Msg("failed to close schema registry")
+		}
 	}
 }