+++ banyand/query/processor.go
@@ -19,6 +19,9 @@ package query
 
 import (
 	"context"
+	"errors"
+	"fmt"
+	"runtime/debug"
 	"time"
 
 	"go.uber.org/multierr"
@@ -34,6 +37,7 @@ import (
 	"github.com/apache/skywalking-banyandb/banyand/stream"
 	"github.com/apache/skywalking-banyandb/pkg/bus"
 	"github.com/apache/skywalking-banyandb/pkg/logger"
+	"github.com/apache/skywalking-banyandb/pkg/query"
 	"github.com/apache/skywalking-banyandb/pkg/query/executor"
 	logical_measure "github.com/apache/skywalking-banyandb/pkg/query/logical/measure"
 	logical_stream "github.com/apache/skywalking-banyandb/pkg/query/logical/stream"
@@ -52,13 +56,13 @@ var (
 )
 
 type queryService struct {
-	log *logger.Logger
-	// TODO: remove the metaService once https://github.com/apache/skywalking/issues/10121 is fixed.
 	metaService metadata.Repo
 	pipeline    queue.Server
+	log         *logger.Logger
 	sqp         *streamQueryProcessor
 	mqp         *measureQueryProcessor
 	tqp         *topNQueryProcessor
+	nodeID      string
 }
 
 type streamQueryProcessor struct {
@@ -76,6 +80,12 @@ func (p *streamQueryProcessor) Rev(message bus.Message) (resp bus.Message) {
 	if p.log.Debug().Enabled() {
 		p.log.Debug().RawJSON("criteria", logger.Proto(queryCriteria)).Msg("received a query request")
 	}
+	defer func() {
+		if err := recover(); err != nil {
+			p.log.Error().Interface("err", err).RawJSON("req", logger.Proto(queryCriteria)).Str("stack", string(debug.Stack())).Msg("panic")
+			resp = bus.NewMessage(bus.MessageID(time.Now().UnixNano()), common.NewError("panic"))
+		}
+	}()
 	// TODO: support multiple groups
 	if len(queryCriteria.Groups) > 1 {
 		resp = bus.NewMessage(bus.MessageID(now), common.NewError("only support one group in the query request"))
@@ -124,11 +134,18 @@ type measureQueryProcessor struct {
 
 func (p *measureQueryProcessor) Rev(message bus.Message) (resp bus.Message) {
 	queryCriteria, ok := message.Data().(*measurev1.QueryRequest)
-	now := time.Now().UnixNano()
+	n := time.Now()
+	now := n.UnixNano()
 	if !ok {
 		resp = bus.NewMessage(bus.MessageID(now), common.NewError("invalid event data type"))
 		return
 	}
+	defer func() {
+		if err := recover(); err != nil {
+			p.log.Error().Interface("err", err).RawJSON("req", logger.Proto(queryCriteria)).Str("stack", string(debug.Stack())).Msg("panic")
+			resp = bus.NewMessage(bus.MessageID(time.Now().UnixNano()), common.NewError("panic"))
+		}
+	}()
 	// TODO: support multiple groups
 	if len(queryCriteria.Groups) > 1 {
 		resp = bus.NewMessage(bus.MessageID(now), common.NewError("only support one group in the query request"))
@@ -160,12 +177,33 @@ func (p *measureQueryProcessor) Rev(message bus.Message) (resp bus.Message) {
 		resp = bus.NewMessage(bus.MessageID(now), common.NewError("fail to analyze the query request for measure %s: %v", meta.GetName(), err))
 		return
 	}
+	ctx := context.Background()
+	var tracer *query.Tracer
+	var span *query.Span
+	if queryCriteria.Trace {
+		tracer, ctx = query.NewTracer(ctx, n.Format(time.RFC3339Nano))
+		span, ctx = tracer.StartSpan(ctx, "data-%s", p.queryService.nodeID)
+		span.Tag("plan", plan.String())
+		defer func() {
+			data := resp.Data()
+			switch d := data.(type) {
+			case *measurev1.QueryResponse:
+				d.Trace = tracer.ToProto()
+			case common.Error:
+				span.Error(errors.New(d.Msg()))
+				resp = bus.NewMessage(bus.MessageID(now), &measurev1.QueryResponse{Trace: tracer.ToProto()})
+			default:
+				panic("unexpected data type")
+			}
+			span.Stop()
+		}()
+	}
 
 	if e := ml.Debug(); e.Enabled() {
 		e.Str("plan", plan.String()).Msg("query plan")
 	}
 
-	mIterator, err := plan.(executor.MeasureExecutable).Execute(executor.WithMeasureExecutionContext(context.Background(), ec))
+	mIterator, err := plan.(executor.MeasureExecutable).Execute(executor.WithMeasureExecutionContext(ctx, ec))
 	if err != nil {
 		ml.Error().Err(err).RawJSON("req", logger.Proto(queryCriteria)).Msg("fail to close the query plan")
 		resp = bus.NewMessage(bus.MessageID(now), common.NewError("fail to execute the query plan for measure %s: %v", meta.GetName(), err))
@@ -174,19 +212,36 @@ func (p *measureQueryProcessor) Rev(message bus.Message) (resp bus.Message) {
 	defer func() {
 		if err = mIterator.Close(); err != nil {
 			ml.Error().Err(err).RawJSON("req", logger.Proto(queryCriteria)).Msg("fail to close the query plan")
+			if span != nil {
+				span.Error(fmt.Errorf("fail to close the query plan: %w", err))
+			}
 		}
 	}()
+
 	result := make([]*measurev1.DataPoint, 0)
-	for mIterator.Next() {
-		current := mIterator.Current()
-		if len(current) > 0 {
-			result = append(result, current[0])
+	func() {
+		var r int
+		if tracer != nil {
+			iterSpan, _ := tracer.StartSpan(ctx, "iterator")
+			defer func() {
+				iterSpan.Tag("rounds", fmt.Sprintf("%d", r))
+				iterSpan.Tag("size", fmt.Sprintf("%d", len(result)))
+				iterSpan.Stop()
+			}()
 		}
-	}
+		for mIterator.Next() {
+			r++
+			current := mIterator.Current()
+			if len(current) > 0 {
+				result = append(result, current[0])
+			}
+		}
+	}()
+	qr := &measurev1.QueryResponse{DataPoints: result}
 	if e := ml.Debug(); e.Enabled() {
-		e.RawJSON("ret", logger.Proto(&measurev1.QueryResponse{DataPoints: result})).Msg("got a measure")
+		e.RawJSON("ret", logger.Proto(qr)).Msg("got a measure")
 	}
-	resp = bus.NewMessage(bus.MessageID(now), &measurev1.QueryResponse{DataPoints: result})
+	resp = bus.NewMessage(bus.MessageID(now), qr)
 	return
 }
 
@@ -194,7 +249,13 @@ func (q *queryService) Name() string {
 	return moduleName
 }
 
-func (q *queryService) PreRun(_ context.Context) error {
+func (q *queryService) PreRun(ctx context.Context) error {
+	val := ctx.Value(common.ContextNodeKey)
+	if val == nil {
+		return errors.New("node id is empty")
+	}
+	node := val.(common.Node)
+	q.nodeID = node.NodeID
 	q.log = logger.GetLogger(moduleName)
 	return multierr.Combine(
 		q.pipeline.Subscribe(data.TopicStreamQuery, q.sqp),