+++ pkg/cmdsetup/liaison.go
@@ -46,13 +46,11 @@ func newLiaisonCmd(runners ...run.Unit) *cobra.Command {
 	}
 	pipeline := pub.New(metaSvc)
 	localPipeline := queue.Local()
-	nodeSel, err := node.NewMaglevSelector()
-	if err != nil {
-		l.Fatal().Err(err).Msg("failed to initiate required node selector")
-	}
-	grpcServer := grpc.NewServer(ctx, pipeline, localPipeline, metaSvc, grpc.NewClusterNodeRegistry(pipeline, nodeSel))
+	nodeSel := node.NewMaglevSelector()
+	nodeRegistry := grpc.NewClusterNodeRegistry(pipeline, nodeSel)
+	grpcServer := grpc.NewServer(ctx, pipeline, localPipeline, metaSvc, nodeRegistry)
 	profSvc := observability.NewProfService()
-	metricSvc := observability.NewMetricService(metaSvc, nil)
+	metricSvc := observability.NewMetricService(metaSvc, pipeline, "liaison", nodeRegistry)
 	httpServer := http.NewServer()
 	dQuery, err := dquery.NewService(metaSvc, localPipeline, pipeline)
 	if err != nil {